DA#229: Modified: Move open Index Info to it's command directory
